Add localized routing layout and redirects
