+++ src/app/[lang]/blog/[slug]/page.tsx
@@ -1,6 +1,6 @@
-// src/app/blog/[slug]/page.tsx
+// src/app/[lang]/blog/[slug]/page.tsx
 
-import type { Metadata } from 'next'
+import type { Metadata } from "next";
 
 import fs from "fs";
 import path from "path";
@@ -65,9 +65,9 @@ export async function generateMetadata({ params }: { params: { slug: string } })
           url: post.image,
           width: 1200,
           height: 630,
-          alt: '',
-        }
-      ]
+          alt: "",
+        },
+      ],
     },
     twitter: {
       images: [
@@ -75,26 +75,24 @@ export async function generateMetadata({ params }: { params: { slug: string } })
           url: post.image,
           width: 800,
           height: 418,
-          alt: '',
-        }
-      ]
+          alt: "",
+        },
+      ],
     },
-    metadataBase: new URL('https://ianaraujo.com')
+    metadataBase: new URL("https://ianaraujo.com"),
   };
 }
 
-const PostPage = async ({ params }: { params: { slug: string } }) => {
-  const { slug } = params;
+const PostPage = async ({ params }: { params: { lang: string; slug: string } }) => {
+  const { lang, slug } = params;
   const post = await getPost(slug);
 
   return (
     <div className="flex justify-center w-full min-h-screen">
       <div className="mt-10 w-full max-w-screen-md px-8 md:px-0">
-        <Header />
+        <Header lang={lang} />
         <div className="flex flex-col space-y-5 mb-10">
-          <p className="w-fit px-2 py-[2px] bg-zinc-200 text-zinc-800 text-sm rounded">
-            {post.tag}
-          </p>
+          <p className="w-fit px-2 py-[2px] bg-zinc-200 text-zinc-800 text-sm rounded">{post.tag}</p>
           <h1 className="text-4xl font-bold leading-tight">{post.title}</h1>
           <div className="flex items-center space-x-4 text-zinc-600">
             <p>{post.date}</p>
