+++ src/app/layout.tsx
@@ -1,19 +1,2 @@
-import "../styles/globals.css";
-import type { Metadata } from "next";
-
-export const metadata: Metadata = {
-  title: {
-    template: "Ian Araujo | %s",
-    default: "Ian Araujo",
-  },
-  description: "Cientista de Dados e IA",
-  metadataBase: new URL('https://ianaraujo.com')
-};
-
-export default function RootLayout({ children }: Readonly<{ children: React.ReactNode }>) {
-  return (
-    <html lang="en">
-      <body>{children}</body>
-    </html>
-  );
-}
+export { dynamicParams, generateMetadata, generateStaticParams } from "./[lang]/layout";
+export { default } from "./[lang]/layout";
